updated file writer test fixtures
