+++ tests/server/test_calling_routes.py
@@ -16,7 +16,7 @@ from ..fixtures import fixture_generic_queue_container
 from ..fixtures import fixture_test_process_manager
 from ..fixtures_process_monitor import fixture_test_monitor
 from ..fixtures_server import fixture_client_and_server_thread_and_shared_values
-from ..fixtures_server import fixture_generic_start_recording_info_in_shared_dict
+from ..fixtures_server import fixture_generic_beta_1_start_recording_info_in_shared_dict
 from ..fixtures_server import fixture_server_thread
 from ..fixtures_server import fixture_test_client
 from ..helpers import put_object_into_queue_and_raise_error_if_eventually_still_empty
@@ -26,7 +26,7 @@ __fixtures__ = [
     fixture_server_thread,
     fixture_test_client,
     fixture_generic_queue_container,
-    fixture_generic_start_recording_info_in_shared_dict,
+    fixture_generic_beta_1_start_recording_info_in_shared_dict,
     fixture_test_monitor,
     fixture_test_process_manager,
 ]
@@ -381,7 +381,7 @@ def test_route_error_message_is_logged(mocker, test_client):
 
 def test_start_recording__returns_no_error_message_with_multiple_hardware_test_recordings(
     test_client,
-    generic_start_recording_info_in_shared_dict,
+    generic_beta_1_start_recording_info_in_shared_dict,
 ):
     response = test_client.get("/start_recording?barcode=MA200440001&is_hardware_test_recording=True")
     assert response.status_code == 200
@@ -390,18 +390,18 @@ def test_start_recording__returns_no_error_message_with_multiple_hardware_test_r
 
 
 def test_start_recording__returns_error_code_and_message_if_user_account_id_not_set(
-    test_client, test_monitor, generic_start_recording_info_in_shared_dict
+    test_client, test_monitor, generic_beta_1_start_recording_info_in_shared_dict
 ):
-    generic_start_recording_info_in_shared_dict["config_settings"]["User Account ID"] = ""
+    generic_beta_1_start_recording_info_in_shared_dict["config_settings"]["User Account ID"] = ""
     response = test_client.get("/start_recording?barcode=MA200440001")
     assert response.status_code == 406
     assert response.status.endswith("User Account ID has not yet been set") is True
 
 
 def test_start_recording__returns_error_code_and_message_if_customer_account_id_not_set(
-    test_client, test_monitor, generic_start_recording_info_in_shared_dict
+    test_client, test_monitor, generic_beta_1_start_recording_info_in_shared_dict
 ):
-    generic_start_recording_info_in_shared_dict["config_settings"]["Customer Account ID"] = ""
+    generic_beta_1_start_recording_info_in_shared_dict["config_settings"]["Customer Account ID"] = ""
     response = test_client.get("/start_recording?barcode=MA200440001")
     assert response.status_code == 406
     assert response.status.endswith("Customer Account ID has not yet been set") is True
@@ -513,7 +513,7 @@ def test_start_recording__allows_years_other_than_20_in_barcode(
     test_client,
     test_barcode,
     test_description,
-    generic_start_recording_info_in_shared_dict,
+    generic_beta_1_start_recording_info_in_shared_dict,
     test_process_manager,
 ):
     response = test_client.get(f"/start_recording?barcode={test_barcode}")
@@ -541,7 +541,7 @@ def test_start_recording__allows_correct_barcode_headers(
     test_barcode,
     test_description,
     test_client,
-    generic_start_recording_info_in_shared_dict,
+    generic_beta_1_start_recording_info_in_shared_dict,
 ):
     response = test_client.get(f"/start_recording?barcode={test_barcode}")
     assert response.status_code == 200
